td-loader: Add fuzz is elf and pe
